WIP refactoring pipeline tests - switching to fast tokenizers
