+++ setup.py
@@ -129,7 +129,7 @@ setup(
         # for OpenAI GPT
         "regex != 2019.12.17",
         # for SentencePiece models
-        "sentencepiece != 0.1.92",
+        # "sentencepiece != 0.1.92",
         "protobuf",
         # for XLM
         "sacremoses",
