+++ src/transformers/tokenization_utils_fast.py
@@ -122,16 +122,11 @@ class PreTrainedTokenizerFast(PreTrainedTokenizerBase):
         return self._tokenizer.get_vocab_size(with_added_tokens=False)
 
     def get_vocab(self) -> Dict[str, int]:
-        """
-        Returns the vocabulary as a dictionary of token to index.
-
-        :obj:`tokenizer.get_vocab()[token]` is equivalent to :obj:`tokenizer.convert_tokens_to_ids(token)` when
-        :obj:`token` is in the vocab.
-
-        Returns:
-            :obj:`Dict[str, int]`: The vocabulary.
-        """
         return self._tokenizer.get_vocab(with_added_tokens=True)
+
+    @property
+    def vocab(self) -> Dict[str, int]:
+        return self.get_vocab()
 
     def get_added_vocab(self) -> Dict[str, int]:
         """
@@ -291,25 +286,8 @@ class PreTrainedTokenizerFast(PreTrainedTokenizerBase):
             tokens.append(self._tokenizer.id_to_token(index))
         return tokens
 
-    def tokenize(self, text: str, pair: Optional[str] = None, add_special_tokens: bool = False) -> List[str]:
-        """
-        Converts a string in a sequence of tokens, using the backend Rust tokenizer.
-
-        Note that, unlike slow tokenizers (instances of :class:`~transformers.PreTrainedTokenizer`), this method
-        will replace the unknown tokens with the :obj:`unk_token`.
-
-        Args:
-            text (:obj:`str`):
-                The sequence to be encoded.
-            pair (:obj:`str`, `optional`):
-                A second sequence to be encoded with the first.
-            add_special_tokens (:obj:`bool`, `optional`, defaults to :obj:`False`):
-                Whether or not to add the special tokens associated with the corresponding model.
-
-        Returns:
-            :obj:`List[str]`: The list of tokens.
-        """
-        return self._tokenizer.encode(text, pair, add_special_tokens=add_special_tokens).tokens
+    def tokenize(self, text: str, pair: Optional[str] = None, add_special_tokens: bool = False, **kwargs) -> List[str]:
+        return self.encode_plus(text=text, text_pair=pair, add_special_tokens=add_special_tokens, **kwargs).tokens()
 
     def set_truncation_and_padding(
         self,
@@ -405,29 +383,11 @@ class PreTrainedTokenizerFast(PreTrainedTokenizerBase):
             pad_to_multiple_of=pad_to_multiple_of,
         )
 
-        # Avoid thread overhead if only one example.
-        if len(batch_text_or_text_pairs) == 1:
-            if isinstance(batch_text_or_text_pairs[0], tuple):
-                # We got a Tuple with a pair of sequences
-                encodings = self._tokenizer.encode(
-                    *batch_text_or_text_pairs[0],
-                    add_special_tokens=add_special_tokens,
-                    is_pretokenized=is_split_into_words,
-                )
-            else:
-                # We got a single sequence
-                encodings = self._tokenizer.encode(
-                    batch_text_or_text_pairs[0],
-                    add_special_tokens=add_special_tokens,
-                    is_pretokenized=is_split_into_words,
-                )
-            encodings = [encodings]
-        else:
-            encodings = self._tokenizer.encode_batch(
-                batch_text_or_text_pairs,
-                add_special_tokens=add_special_tokens,
-                is_pretokenized=is_split_into_words,
-            )
+        encodings = self._tokenizer.encode_batch(
+            batch_text_or_text_pairs,
+            add_special_tokens=add_special_tokens,
+            is_pretokenized=is_split_into_words,
+        )
 
         # Convert encoding to dict
         # `Tokens` has type: List[Dict[str, List[List[int]]]] or List[Dict[str, 2D-Tensor]]
@@ -525,30 +485,16 @@ class PreTrainedTokenizerFast(PreTrainedTokenizerBase):
 
         return batched_output
 
-    def decode(
+    def convert_tokens_to_string(self, tokens: List[str]) -> str:
+        return self.backend_tokenizer.decoder.decode(tokens)
+
+    def _decode(
         self,
         token_ids: Union[int, List[int]],
         skip_special_tokens: bool = False,
         clean_up_tokenization_spaces: bool = True,
         **kwargs
     ) -> str:
-        """
-        Converts a sequence of ids in a string, using the tokenizer and vocabulary
-        with options to remove special tokens and clean up tokenization spaces.
-
-        Similar to doing ``self.convert_tokens_to_string(self.convert_ids_to_tokens(token_ids))``.
-
-        Args:
-            token_ids (:obj:`Union[int, List[int]]`):
-                List of tokenized input ids. Can be obtained using the ``__call__`` method.
-            skip_special_tokens (:obj:`bool`, `optional`, defaults to :obj:`False`):
-                Whether or not to remove special tokens in the decoding.
-            clean_up_tokenization_spaces (:obj:`bool`, `optional`, defaults to :obj:`True`):
-                Whether or not to clean up the tokenization spaces.
-
-        Returns:
-            :obj:`str`: The decoded sentence.
-        """
         if isinstance(token_ids, int):
             token_ids = [token_ids]
         text = self._tokenizer.decode(token_ids, skip_special_tokens=skip_special_tokens)
